app/fastapi/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Demo FastAPI project"
authors = [
    { name = "blueswen", email = "blueswen.tw@gmail.com" }
]
dependencies = [
    "fastapi>=0.114.2",
    "uvicorn>=0.30.6",
    "prometheus-client>=0.20.0",
    "pyroscope-io>=0.8.11",
    "httpx>=0.28.1",
    "opentelemetry-distro>=0.54b1",
    "opentelemetry-exporter-otlp>=1.33.1",
    "opentelemetry-instrumentation-asyncio==0.55b1",
    "opentelemetry-instrumentation-dbapi==0.55b1",
    "opentelemetry-instrumentation-logging==0.55b1",
    "opentelemetry-instrumentation-sqlite3==0.55b1",
    "opentelemetry-instrumentation-threading==0.55b1",
    "opentelemetry-instrumentation-urllib==0.55b1",
    "opentelemetry-instrumentation-wsgi==0.55b1",
    "opentelemetry-instrumentation-click==0.55b1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-instrumentation-grpc==0.55b1",
    "opentelemetry-instrumentation-httpx==0.55b1",
    "opentelemetry-instrumentation-requests==0.55b1",
    "opentelemetry-instrumentation-starlette==0.55b1",
    "opentelemetry-instrumentation-tortoiseorm==0.55b1",
    "opentelemetry-instrumentation-urllib3==0.55b1",
    "psycopg2>=2.9.10",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "opentelemetry-instrumentation-asgi==0.55b1",
    "opentelemetry-instrumentation-psycopg2==0.55b1",
    "opentelemetry-instrumentation-redis==0.55b1",
    "opentelemetry-instrumentation-sqlalchemy==0.55b1",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
